fix: extract camera logic to a separate component and try to make it more bulletproof [PT-185687905]
